moved structure around to fit MSCV model, created necessary service files for custom session handling
