+++ src/app.ts
@@ -16,6 +16,7 @@ import signupRoute from "./routes/signupRoute";
 
 app.set("view engine", "ejs");
 app.use(express.static("public"));
+
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
