+++ src/services/userService.ts
@@ -1,0 +1,7 @@
+import { pool } from "../../database/database";
+import {
+  create_customer,
+  get_customer,
+  get_customers,
+  create_user,
+} from "../models/userModel";
